Refactor project structure: Migrate process_scenario function to utils, remove unused tutorial files
