+++ utils/collect_scenarios.py
